+++ Server_Root/Server/RoomServer/Behavior/IBehavior.h
@@ -22,8 +22,8 @@ protected:
 	bool m_bActivated = false;
 
 protected:
-	__int64	m_lStartTime;			//	ElapsedTime after game was started (Milliseconds)
-	__int64	m_lLastUpdateTime;		//	ElapsedTime after game was started (Milliseconds)
+	long long m_lStartTime;			//	ElapsedTime after game was started (Milliseconds)
+	long long m_lLastUpdateTime;	//	ElapsedTime after game was started (Milliseconds)
 	float m_fDeltaTime;				//	CurrentTime - PreviousTime (seconds)
 	float m_fCurrentTime;			//	ElapsedTime after behavior was started (seconds)
 	float m_fPreviousTime;			//	ElapsedTime after behavior was started (seconds)
@@ -32,15 +32,15 @@ public:
 	int GetMasterDataID();
 
 public:
-	void Update(__int64 lUpdateTime);
-	void Stop(__int64 lTime);
+	void Update(long long lUpdateTime);
+	void Stop(long long lTime);
 
 public:
 	virtual void Initialize() = 0;
-	virtual void UpdateBody(__int64 lUpdateTime) = 0;
+	virtual void UpdateBody(long long lUpdateTime) = 0;
 
 public:
-	virtual void Start(__int64 lStartTime, ...);
+	virtual void Start(long long lStartTime, ...);
 
 public:
 	float GetTime();
